requirements.txt
# EduMindAI Dependencies

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Google Gemini
google-generativeai==0.8.3

# Retry Logic
tenacity==9.0.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.19

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
